requirements.txt
# AI Voice Agent - B2B Lead Generation
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1
aioredis==2.0.1

# AI/ML
openai==1.12.0
anthropic==0.18.0
langchain==0.1.6
langchain-openai==0.0.5
google-generativeai==0.3.2  # Gemini API
google-cloud-aiplatform==1.42.1  # Vertex AI API for GCP VM
llama-cpp-python==0.2.26  # Local LLM runner for GPU

# Speech-to-Text
deepgram-sdk==3.0.0
google-cloud-speech==2.24.0

# Text-to-Speech
elevenlabs==0.3.0
azure-cognitiveservices-speech==1.35.0
edge-tts==6.1.9  # Free high-quality TTS

# Telephony
twilio==8.10.0
# exotel - use requests for API calls

# Web Scraping
playwright==1.41.0
beautifulsoup4==4.12.3
selenium==4.17.0
googlemaps==4.10.0
requests==2.31.0
aiohttp==3.9.3
httpx==0.26.0

# Integrations
hubspot-api-client==8.2.1
google-api-python-client==2.115.0
google-auth-oauthlib==1.2.0
gspread==6.0.0

# WhatsApp
python-whatsapp-business==0.1.0

# Utilities
python-dotenv==1.0.1
pydub==0.25.1
phonenumbers==8.13.28
celery==5.3.6
APScheduler==3.10.4
python-multipart==0.0.7

# Logging & Monitoring
loguru==0.7.2
sentry-sdk==1.39.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0

# Data Processing
pandas==2.2.0
numpy==1.26.3
# ML Auto-Learning System
chromadb>=0.4.22             # Vector database for RAG
sentence-transformers>=2.2.2  # Embeddings for multilingual (Hindi/English)
scikit-learn>=1.3.2          # ML models for intent/lead scoring
torch>=2.1.0                 # Required by sentence-transformers
faiss-cpu>=1.7.4             # Optional: faster vector search